Create solution and update readme

+++ Balance.DataAccess/Dto/TransactionDto.cs
@@ -3,7 +3,7 @@
     public class TransactionDto
     {
         public string Name { get; set; }
-        public decimal Amount { get; set; }
+        public int Amount { get; set; }
         public DateTime CreatedAt { get; set; }
     }
 }
